#42 move persistence classes to own package
