+++ src/main/java/com/leonarduk/itemfinder/freecycle/FreecycleScraper.java
@@ -23,6 +23,7 @@ import org.htmlparser.util.NodeList;
 import org.htmlparser.util.ParserException;
 import org.htmlparser.util.SimpleNodeIterator;
 
+import com.leonarduk.itemfinder.freecycle.db.FreecycleItem;
 import com.leonarduk.itemfinder.html.HtmlParser;
 
 /**
