+++ src/test/java/com/leonarduk/itemfinder/freecycle/FreecycleItemTest.java
@@ -18,6 +18,8 @@ import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 
+import com.leonarduk.itemfinder.freecycle.db.FreecycleItem;
+
 /**
  * The Class FreecycleItemTest.
  *
