+++ src/test/java/com/leonarduk/itemfinder/freecycle/FreecycleScraperTest.java
@@ -20,6 +20,7 @@ import org.junit.Test;
 import org.mockito.Matchers;
 import org.mockito.Mockito;
 
+import com.leonarduk.itemfinder.freecycle.db.FreecycleItem;
 import com.leonarduk.itemfinder.html.HtmlParser;
 
 public class FreecycleScraperTest {
