+++ src/test/java/com/leonarduk/itemfinder/query/QueryReporterTest.java
@@ -11,7 +11,7 @@ import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import com.leonarduk.itemfinder.freecycle.FreecycleItem;
+import com.leonarduk.itemfinder.freecycle.db.FreecycleItem;
 import com.leonarduk.itemfinder.interfaces.Item;
 import com.leonarduk.webscraper.core.format.HtmlFormatter;
